Semana 9 - v4

+++ proyecto/src/main/java/com/proyecto/repository/ClientesRepository.java
@@ -4,7 +4,7 @@ package com.proyecto.repository;
 import com.proyecto.entity.Clientes;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;@Repository
-public interface ClientesRepository extends CrudRepository<Clientes,Long> {
+public interface ClientesRepository extends CrudRepository<Clientes,Integer> {
 }
